Refactor ghost stories: rename runStoryTests→runGhostStories, export from core-server, inline into grade.ts
- Rename runStoryTests to runGhostStories in core (clearer name)
- Add cwd parameter to runGhostStories and getComponentCandidates
- Export getComponentCandidates, runGhostStories, TestRunSummary from core-server index
- Remove eval ghost-stories.ts wrapper — inline logic into grade.ts
- Remove eval ghost-stories.test.ts — core already has its own tests
- Revert speculative isCandidate/isValidCandidate export (unused)
- Remove unused logger import from get-candidates.ts
